# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/branches/{branchId}/carts": {
      "post": {
        "operationId": "CartsController_createCart",
        "parameters": [
          {
            "name": "branchId",
            "required": true,
            "in": "path",
            "description": "Branch id (26 chars)",
            "schema": {
              "example": "01J1QZQ0VQ8J7TQH0YV3A1BCDE",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCartDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Cart created"
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Missing or invalid access token"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create cart (order draft)",
        "tags": [
          "Carts"
        ]
      }
    },
    "/api/v1/branches/{branchId}/carts/{cartId}": {
      "get": {
        "operationId": "CartsController_getCart",
        "parameters": [
          {
            "name": "cartId",
            "required": true,
            "in": "path",
            "description": "Cart id (26 chars)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "branchId",
            "required": true,
            "in": "path",
            "description": "Branch id (26 chars)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Cart detail"
          },
          "401": {
            "description": "Missing or invalid access token"
          },
          "404": {
            "description": "Cart not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get cart by id",
        "tags": [
          "Carts"
        ]
      }
    },
    "/api/v1/branches/{branchId}/carts/{cartId}/items": {
      "post": {
        "operationId": "CartsController_addItem",
        "parameters": [
          {
            "name": "cartId",
            "required": true,
            "in": "path",
            "description": "Cart id (26 chars)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "branchId",
            "required": true,
            "in": "path",
            "description": "Branch id (26 chars)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddCartItemDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Cart updated"
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Missing or invalid access token"
          },
          "404": {
            "description": "Cart not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Add item to cart",
        "tags": [
          "Carts"
        ]
      }
    },
    "/api/v1/branches/{branchId}/carts/{cartId}/items/{productId}": {
      "patch": {
        "operationId": "CartsController_setItemQuantity",
        "parameters": [
          {
            "name": "productId",
            "required": true,
            "in": "path",
            "description": "Product id (26 chars)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cartId",
            "required": true,
            "in": "path",
            "description": "Cart id (26 chars)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "branchId",
            "required": true,
            "in": "path",
            "description": "Branch id (26 chars)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetCartItemQuantityDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Cart updated"
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Missing or invalid access token"
          },
          "404": {
            "description": "Cart not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Set item quantity",
        "tags": [
          "Carts"
        ]
      },
      "delete": {
        "operationId": "CartsController_removeItem",
        "parameters": [
          {
            "name": "productId",
            "required": true,
            "in": "path",
            "description": "Product id (26 chars)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cartId",
            "required": true,
            "in": "path",
            "description": "Cart id (26 chars)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "branchId",
            "required": true,
            "in": "path",
            "description": "Branch id (26 chars)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Cart updated"
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Missing or invalid access token"
          },
          "404": {
            "description": "Cart not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Remove item from cart",
        "tags": [
          "Carts"
        ]
      }
    },
    "/api/v1/branches/{branchId}/carts/{cartId}/checkout": {
      "post": {
        "operationId": "CartsController_checkout",
        "parameters": [
          {
            "name": "cartId",
            "required": true,
            "in": "path",
            "description": "Cart id (26 chars)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "branchId",
            "required": true,
            "in": "path",
            "description": "Branch id (26 chars)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckoutCartDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Checkout result"
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Missing or invalid access token"
          },
          "404": {
            "description": "Cart not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Checkout cart (creates invoice, stock movements)",
        "tags": [
          "Carts"
        ]
      }
    },
...
  },
  "info": {
    "title": "Stock SaaS API",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Send a valid JWT access token. This API also accepts the access token via cookie (see cookie auth)."
      },
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "accessToken",
        "description": "Optional. If present, the API will read the access token from this cookie."
      }
    },
    "schemas": {
...
      "CreateCartDto": {
        "type": "object",
        "properties": {
          "customerId": {
            "type": "string",
            "description": "Optional customer id. Must exist in the same tenant and be active (isActive=true).",
            "minLength": 1
          }
        }
      },
      "AddCartItemDto": {
        "type": "object",
        "properties": {
          "productId": {
            "type": "string",
            "description": "Product id to add (26 chars).",
            "example": "01J1QZQ0VQ8J7TQH0YV3A1BCDE"
          },
          "quantity": {
            "type": "number",
            "description": "Quantity to add (>= 1).",
            "example": 1,
            "minimum": 1
          }
        },
        "required": [
          "productId",
          "quantity"
        ]
      },
      "SetCartItemQuantityDto": {
        "type": "object",
        "properties": {
          "quantity": {
            "type": "number",
            "description": "New quantity. Use 0 to remove the item.",
            "example": 2,
            "minimum": 0
          }
        },
        "required": [
          "quantity"
        ]
      },
      "CheckoutCartDto": {
        "type": "object",
        "properties": {
          "customerId": {
            "type": "string",
            "description": "Optional customer id to (re)assign on checkout. Must exist in the same tenant and be active (isActive=true).",
            "minLength": 1
          }
        }
      },
...
    }
  }
}